Create separate Serve- and Get- functions
Get- functions deal with retrieving stuff from the backend and returning
useful objects.

Serve- functions deal with getting the relevant params from the query,
invoking the relevant Get- functions, and serving up the objects as
JSON.

Why we do this:
1. Objects returned from Get- functions could be used elsewhere (search
maybe, definitely in building other objects)

2. Allows for cleaner code. Much of the ugly error handling was
refactored in this process because Get- functions can return error
objects of their own now instead of having to handle it themselves, and
so any errors in a Get- function is now dealt with in 3 lines of the
Serve- function. I also was able to create a separate Serve() helper
function in router.go that deals with writing any Go object to a
http.ResponseWriter.